# Conflicts:
#	code/modules/projectiles/gun_system.dm

+++ code/modules/projectiles/projectile.dm
@@ -61,6 +61,10 @@
 	var/list/atom/movable/uncross_scheduled = list() // List of border movable atoms to check for when exiting a turf.
 
 	var/damage = 0
+	///ammo penetration value
+	var/penetration = 0
+	///ammo sundering value
+	var/sundering = 0
 	var/accuracy = 85 //Base projectile accuracy. Can maybe be later taken from the mob if desired.
 
 	///how many damage points the projectile loses per tiles travelled
@@ -121,8 +125,9 @@
 		return
 	if(AM.projectile_hit(src))
 		AM.do_projectile_hit(src)
-		qdel(src)
-		return
+		if(!(ammo.flags_ammo_behavior & AMMO_PASS_THROUGH_MOVABLE))
+			qdel(src)
+			return
 	permutated[AM] = TRUE //Don't want to hit them again.
 
 
@@ -138,6 +143,8 @@
 	name 		= ammo.name
 	icon_state 	= ammo.icon_state
 	damage 		= ammo.damage + bonus_damage //Mainly for emitters.
+	penetration = ammo.penetration
+	sundering 	= ammo.sundering
 	scatter		= ammo.scatter
 	accuracy   += ammo.accuracy
 	accuracy   *= rand(95 - ammo.accuracy_var_low, 105 + ammo.accuracy_var_high) * 0.01 //Rand only works with integers.
@@ -434,6 +441,8 @@ So if we are on the 32th absolute pixel coordinate we are on tile 1, but if we a
 				if(!thing_to_uncross.projectile_hit(src, REVERSE_DIR(border_escaped_through), TRUE))
 					continue
 				thing_to_uncross.do_projectile_hit(src)
+				if(ammo.flags_ammo_behavior & AMMO_PASS_THROUGH_MOVABLE)
+					continue
 				end_of_movement = i
 				break
 			uncross_scheduled.len = 0
@@ -479,6 +488,8 @@ So if we are on the 32th absolute pixel coordinate we are on tile 1, but if we a
 				if(!thing_to_uncross.projectile_hit(src, REVERSE_DIR(movement_dir), TRUE))
 					continue
 				thing_to_uncross.do_projectile_hit(src)
+				if(ammo.flags_ammo_behavior & AMMO_PASS_THROUGH_MOVABLE)
+					continue
 				end_of_movement = i
 				break
 			uncross_scheduled.len = 0
@@ -500,6 +511,8 @@ So if we are on the 32th absolute pixel coordinate we are on tile 1, but if we a
 				if(!thing_to_uncross.projectile_hit(src, REVERSE_DIR(movement_dir), TRUE))
 					continue //We act as if we were entering the tile through the opposite direction, to check for barricade blockage.
 				thing_to_uncross.do_projectile_hit(src)
+				if(ammo.flags_ammo_behavior & AMMO_PASS_THROUGH_MOVABLE)
+					continue
 				end_of_movement = i
 				break
 			uncross_scheduled.len = 0
@@ -557,7 +570,7 @@ So if we are on the 32th absolute pixel coordinate we are on tile 1, but if we a
 	if(turf_to_scan.density) //Handle wall hit.
 		ammo.on_hit_turf(turf_to_scan, src)
 		turf_to_scan.bullet_act(src)
-		return TRUE
+		return !(ammo.flags_ammo_behavior & AMMO_PASS_THROUGH_TURF)
 
 	if(shot_from)
 		switch(SEND_SIGNAL(shot_from, COMSIG_PROJ_SCANTURF, turf_to_scan))
@@ -578,7 +591,7 @@ So if we are on the 32th absolute pixel coordinate we are on tile 1, but if we a
 			continue
 
 		thing_to_hit.do_projectile_hit(src)
-		return TRUE
+		return !(ammo.flags_ammo_behavior & AMMO_PASS_THROUGH_MOVABLE)
 
 	return FALSE
 
@@ -822,6 +835,11 @@ So if we are on the 32th absolute pixel coordinate we are on tile 1, but if we a
 //----------------------------------------------------------
 
 /atom/proc/bullet_act(obj/projectile/proj)
+	if(HAS_TRAIT(proj, TRAIT_PROJ_HIT_SOMETHING))
+		proj.damage *= proj.ammo.on_pierce_multiplier
+		proj.penetration *= proj.ammo.on_pierce_multiplier
+		proj.sundering *= proj.ammo.on_pierce_multiplier
+	ADD_TRAIT(proj, TRAIT_PROJ_HIT_SOMETHING, BULLET_ACT_TRAIT)
 	SEND_SIGNAL(src, COMSIG_ATOM_BULLET_ACT, proj)
 
 
@@ -851,16 +869,15 @@ So if we are on the 32th absolute pixel coordinate we are on tile 1, but if we a
 	var/living_hard_armor = (proj.ammo.flags_ammo_behavior & AMMO_IGNORE_ARMOR) ? 0 : get_hard_armor(proj.armor_type, proj.def_zone, proj.dir)
 	var/living_soft_armor = (proj.ammo.flags_ammo_behavior & AMMO_IGNORE_ARMOR) ? 0 : get_soft_armor(proj.armor_type, proj.def_zone, proj.dir)
 	if(living_hard_armor || living_soft_armor)
-		var/penetration = proj.ammo.penetration
-		if(penetration > 0)
+		if(proj.penetration > 0)
 			if(proj.shot_from && src == proj.shot_from.sniper_target(src))
 				damage *= SNIPER_LASER_DAMAGE_MULTIPLIER
-				penetration *= SNIPER_LASER_ARMOR_MULTIPLIER
+				proj.penetration *= SNIPER_LASER_ARMOR_MULTIPLIER
 				add_slowdown(SNIPER_LASER_SLOWDOWN_STACKS)
 			if(living_hard_armor)
-				living_hard_armor = max(0, living_hard_armor - (living_hard_armor * penetration * 0.01)) //AP reduces a % of hard armor.
+				living_hard_armor = max(0, living_hard_armor - (living_hard_armor * proj.penetration * 0.01)) //AP reduces a % of hard armor.
 			if(living_soft_armor)
-				living_soft_armor = max(0, living_soft_armor - penetration) //Flat removal.
+				living_soft_armor = max(0, living_soft_armor - proj.penetration) //Flat removal.
 
 		if(iscarbon(proj.firer))
 			var/mob/living/carbon/shooter_carbon = proj.firer
@@ -892,7 +909,7 @@ So if we are on the 32th absolute pixel coordinate we are on tile 1, but if we a
 			feedback_flags |= (BULLET_FEEDBACK_FIRE|BULLET_FEEDBACK_SCREAM)
 
 	if(proj.ammo.flags_ammo_behavior & AMMO_SUNDERING)
-		adjust_sunder(proj.ammo.sundering)
+		adjust_sunder(proj.sundering)
 
 	if(damage)
 		var/shrapnel_roll = do_shrapnel_roll(proj, damage)
@@ -1041,6 +1058,8 @@ So if we are on the 32th absolute pixel coordinate we are on tile 1, but if we a
 				if(!thing_to_uncross.projectile_hit(src, REVERSE_DIR(border_escaped_through), TRUE))
 					continue
 				thing_to_uncross.do_projectile_hit(src)
+				if(ammo.flags_ammo_behavior & AMMO_PASS_THROUGH_MOVABLE)
+					continue
 				end_of_movement = TRUE
 				break
 			uncross_scheduled.Cut()
@@ -1063,6 +1082,8 @@ So if we are on the 32th absolute pixel coordinate we are on tile 1, but if we a
 				if(!thing_to_uncross.projectile_hit(src, REVERSE_DIR(movement_dir), TRUE))
 					continue
 				thing_to_uncross.do_projectile_hit(src)
+				if(ammo.flags_ammo_behavior & AMMO_PASS_THROUGH_MOVABLE)
+					continue
 				end_of_movement = TRUE
 				break
 			uncross_scheduled.Cut()
@@ -1077,6 +1098,8 @@ So if we are on the 32th absolute pixel coordinate we are on tile 1, but if we a
 				if(!thing_to_uncross.projectile_hit(src, REVERSE_DIR(movement_dir), TRUE))
 					continue //We act as if we were entering the tile through the opposite direction, to check for barricade blockage.
 				thing_to_uncross.do_projectile_hit(src)
+				if(ammo.flags_ammo_behavior & AMMO_PASS_THROUGH_MOVABLE)
+					continue
 				end_of_movement = TRUE
 				break
 			uncross_scheduled.len = 0
@@ -1200,6 +1223,7 @@ So if we are on the 32th absolute pixel coordinate we are on tile 1, but if we a
 
 //Turf handling.
 /turf/bullet_act(obj/projectile/proj)
+	. = ..()
 	bullet_ping(proj)
 
 	var/list/mob_list = list() //Let's built a list of mobs on the bullet turf and grab one.
